+++ src/lib/common/constants.ts
@@ -1,4 +1,49 @@
-import {GoogleMapLibrary} from './types'
+import {GoogleMapLibrary, HeatmapLayerOptions} from './types'
+
+export const NYC_LATLNG: google.maps.LatLngLiteral = {
+  lat: 40.7128,
+  lng: -74.006,
+}
+const NYC_POLYGON: google.maps.LatLngLiteral[] = [
+  {lat: 40.718, lng: -74.006},
+  {lat: 40.718, lng: -74.01},
+  {lat: 40.7138, lng: -74.001},
+  {lat: 40.7128, lng: -74.005},
+]
+
+const DEFAULT_MAP_ZOOM = 14
+
+export const DEFAULT_HEAT_MAP_OPTIONS: HeatmapLayerOptions = {
+  data: [{lat: 40.718, lng: -74.006}, {lat: 40.712, lng: -74.01}],
+  radius: 100,
+}
+
+export const DEFAULT_INFO_WINDOW_OPTIONS: google.maps.InfoWindowOptions = {
+  content: 'Hello',
+  position: NYC_LATLNG,
+}
+
+export const DEFAULT_MAP_OPTIONS: google.maps.MapOptions = {
+  center: NYC_LATLNG,
+  zoom: DEFAULT_MAP_ZOOM,
+}
+
+export const DEFAULT_MAP_STYLE = {
+  height: '100%',
+  width: '100%',
+}
+
+export const DEFAULT_MARKER_OPTIONS: google.maps.MarkerOptions = {
+  position: NYC_LATLNG,
+}
+
+export const DEFAULT_POLYGON_OPTIONS: google.maps.PolygonOptions = {
+  paths: NYC_POLYGON,
+}
+
+export const DEFAULT_STREET_VIEW_OPTIONS: google.maps.StreetViewPanoramaOptions = {
+  position: NYC_LATLNG,
+}
 
 export const GOOGLE_MAP_BASE_URI = 'https://maps.googleapis.com/maps/api/js'